TASK 1
# Guess the Number Game Project

## Description
The "Guess the Number" game project is an interactive Java program where the user tries to guess a randomly generated number. The program provides feedback to the user on whether the guessed number is greater, lower, or the correct number. Additionally, it keeps track of the number of attempts made by the user.

## Game Flow
1. The program generates a random number.
2. The user inputs a guess.
3. If the guess is incorrect, the program informs the user if the number is greater or lower than the guessed number.
4. The user continues to guess until they correctly identify the random number or reach the maximum number of attempts.
5. The program ends if the user guesses the correct number within 10 attempts or exceeds the maximum attempts allowed.

## Number of Attempts
The game limits the user to 10 attempts to guess the correct number. If the user exceeds this limit, the game terminates automatically.

## Replay Option
After each game, the program prompts the user if they want to play again. If the user chooses to continue, a new random number is generated for another round of the game.