TASK 3
Title: ATM Interface Program

Description:
The ATM Interface Program is a Java project designed to simulate an Automated Teller Machine (ATM) experience. When a user runs the program, they are prompted with options to choose from:
1. Withdrawal: Allows the user to withdraw a certain amount from their account.
2. Deposit: Enables the user to deposit funds into their account.
3. Check Balance: Displays the current balance in the user's account.
4. Exit: Exits the program.

Based on the user's choice, the program performs the corresponding action. For example, if the user selects withdrawal, they will be asked to enter the amount they wish to withdraw, and the program will deduct that amount from their account balance. Similarly, for deposit, the program will prompt the user to input the amount to be deposited and update the account balance accordingly. Checking the balance will simply display the current balance without any changes.

This project showcases the use of conditional statements and user input handling in Java to create an interactive ATM-like experience. It provides a practical demonstration of how programming can be used to simulate real-world scenarios.